Include jquery cookie for accessing login info.
Going to implement a client-server key based authentication as well.
When the user logs in they will always have to send the key the server
replies with to do actions that require login. The key is session
specific, meaning it validates against the IP and whatever else I decide
to configure it to validate against.